+++ ChangeLog.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Formula, incolour, outcolour and plane
+
 ## [0.7.0] 2018-02-20
 
 ### Added